Refactor election signal handling to generate encryption keys on creation
This commit updates the `AppConfig` class to import the new `election_signals` module, which contains a signal receiver for the `post_save` event of the `Election` model. When a new election instance is created, it automatically generates and assigns public and private encryption keys to the instance. Additionally, a new `__init__.py` file is added to the `signals` directory to facilitate module imports.